[Launchpad] Integrate TreeView with new mock data fitting our structure.
This commit has a lot of changes on a lot of files because this is the first time integrating real launchpad session into nearly all components and sections. We may need to deal with some conflicts, but in the future we are not going to deal with them again and we are good to move forward now!